Added type conversion overloads for Fp32f to float. Added appropriate unit tests. Added size_t cast to integers in FpTest-FpxxfSize.cpp.

+++ test/FpTest-Fp32fSize.cpp
@@ -22,6 +22,6 @@ SUITE(Fp32fSizeTests)
 		//return UnitTest::RunAllTests();
 		printf("FP: Size of Fp32f object (bytes) = '%lu'.\r\n", sizeof(Fp32f<12>));
 	
-		CHECK_EQUAL(4, sizeof(Fp32f<12>));
+		CHECK_EQUAL((size_t)4, sizeof(Fp32f<12>));
 	}
 }
